+++ CHANGES.rst
@@ -1,10 +1,13 @@
 Changelog
 =========
 
+There's a frood who really knows where his towel is.
+
 1.0b1 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Add code validation for control panel configlet fields (closes `#1`_).
+  [rodfersou]
 
 
 1.0a2 (2015-04-30)
@@ -21,3 +24,5 @@ Changelog
 ------------------
 
 - Initial release.
+
+.. _`#1`: https://github.com/simplesconsultoria/collective.blueline/issues/1